proxy-admin/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
